requirements.txt
arrow==1.2.0
asgiref==3.4.1
binaryornot==0.4.4
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-organizations==2.0.0
django-sslserver==0.22
django-storages==1.12
django-tailwind==2.2.1
fake-factory==9999.9.9
Faker==8.11.0
firebase-admin==5.0.3
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-secret-manager==2.7.2
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
gunicorn==20.1.0
httplib2==0.20.1
idna==3.2
itypes==1.2.0
Jinja2==3.0.2
jinja2-time==0.2.0
libcst==0.3.21
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.0
Pillow==8.3.1
poyo==0.5.0
proto-plus==1.19.3
protobuf==3.18.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.4.1
qrcode==7.2
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
typing-inspect==0.7.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.7
whitenoise==5.3.0
